[Trainings] fixes course edition access
