+++ src/plugin/cursus/Resources/modules/course/components/main.jsx
@@ -3,10 +3,14 @@ import {PropTypes as T} from 'prop-types'
 import get from 'lodash/get'
 
 import {Routes} from '#/main/app/router/components/routes'
-
 import {route} from '#/plugin/cursus/routing'
+import {hasPermission} from '#/main/app/security'
 import {Course as CourseTypes, Session as SessionTypes} from '#/plugin/cursus/prop-types'
 import {CourseDetails} from '#/plugin/cursus/course/components/details'
+import {CourseForm} from '#/plugin/cursus/course/containers/form'
+
+// should be moved inside the current module instead
+import {selectors} from '#/plugin/cursus/tools/trainings/catalog/store'
 
 const CourseMain = (props) =>
   <Routes
@@ -28,6 +32,16 @@ const CourseMain = (props) =>
             activeSessionRegistration={null}
             availableSessions={props.availableSessions}
             courseRegistration={props.courseRegistration}
+          />
+        )
+      }, {
+        path: '/edit',
+        onEnter: () => props.openForm(props.course.slug),
+        disabled: !hasPermission('edit', props.course),
+        render: () => (
+          <CourseForm
+            path={props.path}
+            name={selectors.FORM_NAME}
           />
         )
       }, {
@@ -65,7 +79,8 @@ CourseMain.propTypes = {
     // TODO : propTypes
   }),
   courseRegistration: T.shape({}),
-  openSession: T.func.isRequired
+  openSession: T.func.isRequired,
+  openForm: T.func.isRequired
 }
 
 export {
